requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.28.2
openai==1.3.3
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
pypdf==3.17.1